+++ include/svs/index/flat/flat.h
@@ -71,6 +71,42 @@ data::GetDatumAccessor svs_invoke(svs::tag_t<accessor>, const Data& SVS_UNUSED(d
     return data::GetDatumAccessor{};
 }
 
+/////
+///// Distance
+/////
+struct ComputeDistanceType {
+    template <typename Data, typename Distance, typename Query>
+    double operator()(
+        const Data& data, const Distance& distance, size_t internal_id, const Query& query
+    ) const {
+        return svs_invoke(*this, data, distance, internal_id, query);
+    }
+};
+// CPO for distance computation
+inline constexpr ComputeDistanceType get_distance_ext{};
+template <typename Data, typename Distance, typename Query>
+double svs_invoke(
+    svs::tag_t<get_distance_ext>,
+    const Data& data,
+    const Distance& distance,
+    size_t internal_id,
+    const Query& query
+) {
+    // Convert query to span for uniform handling
+    auto query_span = lib::as_const_span(query);
+
+    // Get distance
+    auto dist_f = extensions::distance(data, distance);
+    svs::distance::maybe_fix_argument(dist_f, query_span);
+
+    // Get the vector from the index
+    auto indexed_span = data.get_datum(internal_id);
+
+    // Compute the distance using the appropriate distance function
+    auto dist = svs::distance::compute(dist_f, query_span, indexed_span);
+
+    return static_cast<double>(dist);
+}
 } // namespace extensions
 
 // The flat index is "special" because we wish to enable the `FlatIndex` to either:
